# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_gopal-chhetri_url-shortener_internal_response.PaginatedResponse:
    properties:
      data: {}
      error:
        type: boolean
      message:
        type: string
      metadata:
        $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_utils.PaginationResponse'
    type: object
  github_com_gopal-chhetri_url-shortener_internal_response.Response:
    properties:
      data: {}
      error:
        type: boolean
      message:
        type: string
    type: object
  github_com_gopal-chhetri_url-shortener_internal_utils.PaginationResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  internal_url.CreateURLRequestBody:
    properties:
      custom_slug:
        example: my-custom-slug
        type: string
      original_url:
        example: https://example.com/very-long-url
        type: string
    required:
    - original_url
    type: object
  internal_url.URLResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      original_url:
        type: string
      short_url:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  internal_url.UpdateURLRequestBody:
    properties:
      original_url:
        example: https://example.com/new-url
        type: string
    required:
    - original_url
    type: object
host: localhost:8000
info:
  contact: {}
  description: This is a sample URL Shortener server with authentication and authorization.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: URL Shortener API
  version: "1.0"
paths:
...
  /urls:
    get:
      description: Get a paginated list of URLs created by the authenticated user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_url.URLResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
      security:
      - BearerAuth: []
      summary: List user's URLs
      tags:
      - urls
    post:
      consumes:
      - application/json
      description: Create a new shortened URL from an original URL
      parameters:
      - description: Create URL request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_url.CreateURLRequestBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_url.URLResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
      security:
      - BearerAuth: []
      summary: Create a short URL
      tags:
      - urls
  /urls/{id}:
    delete:
      description: Soft delete a URL (deactivate it)
      parameters:
      - description: URL ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
      security:
      - BearerAuth: []
      summary: Delete a URL
      tags:
      - urls
    get:
      description: Get a specific URL by its ID
      parameters:
      - description: URL ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_url.URLResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
      security:
      - BearerAuth: []
      summary: Get URL by ID
      tags:
      - urls
    put:
      consumes:
      - application/json
      description: Update the original URL for an existing short URL
      parameters:
      - description: URL ID
        in: path
        name: id
        required: true
        type: string
      - description: Update URL request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_url.UpdateURLRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_url.URLResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
      security:
      - BearerAuth: []
      summary: Update a URL
      tags:
      - urls
  /urls/{id}/analytics:
    get:
      description: Returns daily clicks (past 7 days), device breakdown, and browser
        breakdown
      parameters:
      - description: URL ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
      security:
      - BearerAuth: []
      summary: Get URL analytics
      tags:
      - urls
  /urls/{id}/status:
    patch:
      consumes:
      - application/json
      description: Activate or deactivate a URL (soft enable/disable)
      parameters:
      - description: URL ID
        in: path
        name: id
        required: true
        type: string
      - description: Status update
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_url.URLResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
      security:
      - BearerAuth: []
      summary: Toggle URL active status
      tags:
      - urls
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"